xfree86: fix obscure bug in xf86platformProbeDev()'s retval computation
According to the leading comment on, and the implementation of,
xf86CallDriverProbe(), the xf86platformProbeDev() function is supposed to
return TRUE if the driver in question succeeds to probe at least one
graphics card.

The assignment to "foundScreen" in the "devList" loop of
xf86platformProbeDev() doesn't calculate a logical disjunction however; a
later probing failure currently suppresses earlier successes. Fix it.

(The assignment to "foundScreen" in the "autoAddGPU" loop is correct
already, no TRUE -> FALSE transition is possible there.)

In addition, remove the useless "continue" statement at the end of the
"devList" loop body.

Cc: Adam Jackson <ajax@redhat.com>
Cc: Dave Airlie <airlied@redhat.com>
Cc: Keith Packard <keithp@keithp.com>
Signed-off-by: Laszlo Ersek <lersek@redhat.com>